Refactor TerminalInfo and CellSize properties to use set accessors; update documentation for ConvertTo-Sixel and ConvertTo-SixelGif cmdlets to include Stream parameter; add AssemblyResolver for net472 compatibility; improve build script and add ShowSixelGif cmdlet.
