+++ module/Sixel.psm1
@@ -1,19 +1,35 @@
 $importModule = Get-Command -Name Import-Module -Module Microsoft.PowerShell.Core
 $moduleName = [System.IO.Path]::GetFileNameWithoutExtension($PSCommandPath)
 
-# This is used to load the shared assembly in the Default ALC which then sets
-# an ALC for the module and any dependencies of that module to be loaded in
-# that ALC.
+if (-not $IsCoreClr) {
+    # PowerShell 5.1 has no concept of an Assembly Load Context so it will
+    # just load the module assembly directly.
 
-$isReload = $true
-if (-not ('Sixel.Shared.LoadContext' -as [type])) {
-    $isReload = $false
-    Add-Type -Path ([System.IO.Path]::Combine($PSScriptRoot, 'bin', 'net8.0', "$moduleName.Shared.dll"))
+    # The type can be any type within our ALCLoader project
+    $innerMod = if ('Sixel.Terminal.SizeHelper' -as [type]) {
+        $modAssembly = [Sixel.Terminal.SizeHelper].Assembly
+        & $importModule -Assembly $modAssembly -Force -PassThru
+    }
+    else {
+        $modPath = [System.IO.Path]::Combine($PSScriptRoot, 'bin', 'net472', "$moduleName.dll")
+        & $importModule -Name $modPath -ErrorAction Stop -PassThru
+    }
 }
+else {
+    # This is used to load the shared assembly in the Default ALC which then sets
+    # an ALC for the module and any dependencies of that module to be loaded in
+    # that ALC.
 
-$mainModule = [Sixel.Shared.LoadContext]::Initialize()
-$innerMod = & $importModule -Assembly $mainModule -PassThru:$isReload
+    $isReload = $true
+    if (-not ('Sixel.Shared.LoadContext' -as [type])) {
+        $isReload = $false
 
+        Add-Type -Path ([System.IO.Path]::Combine($PSScriptRoot, 'bin', 'net8.0', "$moduleName.Shared.dll"))
+    }
+
+    $mainModule = [Sixel.Shared.LoadContext]::Initialize()
+    $innerMod = & $importModule -Assembly $mainModule -PassThru:$isReload
+}
 
 if ($innerMod) {
     # Bug in pwsh, Import-Module in an assembly will pick up a cached instance
