+++ src/Sixel.Shared/SharedUtil.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
+#if NET5_0_OR_GREATER
 using System.Runtime.Loader;
+#endif
 
 namespace Sixel.Shared;
 
@@ -14,7 +16,9 @@ internal class SharedUtil
     data["Assembly"] = new Dictionary<string, object?>()
         {
             { "Name", asm.GetName().FullName },
+#if NET5_0_OR_GREATER
             { "ALC", AssemblyLoadContext.GetLoadContext(asm)?.Name },
+#endif
             { "Location", asm.Location }
         };
   }
