+++ src/Sixel/Cmdlet/ConvertToSixel.cs
@@ -1,19 +1,19 @@
-﻿using Sixel.Terminal;
-using Sixel.Terminal.Models;
-using Sixel.Protocols;
 using System.Management.Automation;
-using System.Threading;
+using Sixel.Terminal;
+using Sixel.Terminal.Models;
+using System.Net.Http;
 
 namespace Sixel.Cmdlet;
 
-[Cmdlet(VerbsData.ConvertTo, "SixelGif", DefaultParameterSetName = "Path")]
-[Alias("gif")]
-[OutputType(typeof(SixelGif))]
-public sealed class ConvertSixelGifCmdlet : PSCmdlet
+[Cmdlet(VerbsData.ConvertTo, "Sixel", DefaultParameterSetName = "Path")]
+[Alias("cts", "ConvertTo-InlineImage","ConvertTo-KittyImage")]
+[OutputType(typeof(string))]
+public sealed class ConvertSixelCmdlet : PSCmdlet
 {
   [Parameter(
         HelpMessage = "A path to a local image to convert to sixel.",
         Mandatory = true,
+        ValueFromPipeline = true,
         ValueFromPipelineByPropertyName = true,
         Position = 0,
         ParameterSetName = "Path"
@@ -26,11 +26,24 @@ public sealed class ConvertSixelGifCmdlet : PSCmdlet
         HelpMessage = "A URL of the image to download and convert to sixel.",
         Mandatory = true,
         ValueFromPipeline = true,
+        ValueFromPipelineByPropertyName = true,
         ParameterSetName = "Url"
   )]
   [ValidateNotNullOrEmpty]
   [Alias("Uri")]
-  public string Url { get; set; } = null!;
+  public Uri Url { get; set; } = null!;
+
+  [Parameter(
+        HelpMessage = "A stream of the image to convert to sixel.",
+        Mandatory = true,
+        ValueFromPipeline = true,
+        ValueFromPipelineByPropertyName = true,
+        Position = 0,
+        ParameterSetName = "Stream"
+  )]
+  [ValidateNotNullOrEmpty]
+  [Alias("FileStream", "InputStream", "ImageStream", "ContentStream")]
+  public Stream Stream { get; set; } = null!;
 
   [Parameter(
         HelpMessage = "The maximum number of colors to use in the image."
@@ -50,14 +63,10 @@ public sealed class ConvertSixelGifCmdlet : PSCmdlet
   public SwitchParameter Force { get; set; }
 
   [Parameter(
-      HelpMessage = "The number of times to loop the gif."
+        HelpMessage = "Choose ImageProtocol to use for conversion."
   )]
-  [ValidateRange(1, 256)]
-  public int LoopCount { get; set; } = 3;
-  [Parameter(
-        HelpMessage = "The audio track to overlay the gif"
-  )]
-  public string? AudioFile { get; set; }
+  public ImageProtocol Protocol { get; set; } = Compatibility.GetTerminalInfo().Protocol;
+
   protected override void ProcessRecord()
   {
     try
@@ -76,56 +85,32 @@ public sealed class ConvertSixelGifCmdlet : PSCmdlet
             using var client = new HttpClient();
             var response = client.GetAsync(Url).Result;
             response.EnsureSuccessStatusCode();
-            imageStream = response.Content.ReadAsStream();
+            // imageStream = response.Content.ReadAsStream();
+            imageStream = response.Content.ReadAsStreamAsync().Result;
+            break;
+          }
+        case "Stream":
+          {
+            if (Stream.Position != 0)
+            {
+              // if something has already read the stream, reset it.. maybe risky
+              Stream.Position = 0;
+            }
+            imageStream = Stream;
             break;
           }
       }
-
       if (imageStream is null) return;
-
       using (imageStream)
       {
-        if (AudioFile is not null)
-        {
-          var resolvedAudio = SessionState.Path.GetResolvedPSPathFromPSPath(AudioFile)[0].Path;
-          WriteObject(GifToSixel.LoadGif(imageStream, MaxColors, Width, LoopCount, resolvedAudio));
-        }
-        else {
-          WriteObject(GifToSixel.LoadGif(imageStream, MaxColors, Width, LoopCount));
-        }
+        WriteObject(
+        Load.ConsoleImage(
+              Protocol,
+              imageStream,
+              MaxColors,
+              Width,
+              Force.IsPresent));
       }
-    }
-    catch (Exception ex)
-    {
-      WriteError(new ErrorRecord(ex, "SixelError", ErrorCategory.NotSpecified, null));
-    }
-  }
-}
-
-[Cmdlet(VerbsCommon.Show, "SixelGif", DefaultParameterSetName = "Path")]
-public sealed class ShowSixelGifCmdlet : PSCmdlet
-{
-  [Parameter(
-    HelpMessage = "SixelGif object to play.",
-    Mandatory = true,
-    ValueFromPipeline = true,
-    Position = 0
-  )]
-  [ValidateNotNullOrEmpty]
-  public SixelGif? Gif { get; set; }
-  protected override void ProcessRecord()
-  {
-    try
-    {
-      if (Gif is null) return;
-      CancellationTokenSource cts = new CancellationTokenSource();
-      // Handle Ctrl+C
-      Console.CancelKeyPress += (sender, args) =>
-      {
-        args.Cancel = true;
-        cts.Cancel();
-      };
-        GifToSixel.PlaySixelGif(Gif, cts.Token);
     }
     catch (Exception ex)
     {
