+++ src/Sixel/Terminal/Models/TerminalInfo.cs
@@ -2,6 +2,6 @@
 
 public class TerminalInfo
 {
-  public Terminals Terminal { get; init; }
-  public ImageProtocol Protocol { get; init; }
+  public Terminals Terminal { get; set; }
+  public ImageProtocol Protocol { get; set; }
 }
